fix the editable projection
Problem:
Scrip error on the 2nd rendering.
In the 2nd rendering of the grid, the editable projection is trying to add the 'grid-editable-cell' class to the column again regardless the existing classes, and the class attribute in the column metadata is turned into a string, no longer an array.

Fix:
1. Accept a string value from $metadata.attrBody.class, split it
2. Use _.union to add the 'grid-editable-cell' class to remove duplications.