Add a few new APIs
1. MemoryDataSource#cachedItems
This is a property to get an array of items with current filter and
order, without skip and take. This can be cached when people are
navigating cross multiple pages. It's not necessary to repeat the
filtering and orderring.

2. GridView#itemOfElement(el)
Combination of GridView#keyOfElement(el) and GridView#itemWithKey(key).
People always use these 2 methods together. It's more convenient to
provide a itemOfElement(el).