Fix DataSource.query getting called unnecessarily
If DataSource object is passed in through config.dataSource, which is a supported scenario, this will expose its query method in config.query.query. Plugins/other consumers expect this to be the query configuration and not the DataSource object. This can cause multiple calls if some plugin calls _.result(config.query, 'query') in order to retrieve this.